+++ src/main/java/net/adventurez/entity/DubiousPumpkinEntity.java
@@ -1,0 +1,192 @@
+package net.adventurez.entity;
+
+import net.adventurez.init.SoundInit;
+import net.minecraft.block.BlockState;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.ai.goal.*;
+import net.minecraft.entity.attribute.DefaultAttributeContainer;
+import net.minecraft.entity.attribute.EntityAttributes;
+import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.data.DataTracker;
+import net.minecraft.entity.data.TrackedData;
+import net.minecraft.entity.data.TrackedDataHandlerRegistry;
+import net.minecraft.entity.mob.Angerable;
+import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.entity.mob.PathAwareEntity;
+import net.minecraft.entity.mob.ZombieEntity;
+import net.minecraft.entity.passive.FoxEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.sound.SoundEvent;
+import net.minecraft.util.TimeHelper;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.intprovider.UniformIntProvider;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.UUID;
+
+public class DubiousPumpkinEntity extends PathAwareEntity implements Angerable {
+    public static final TrackedData<Boolean> ANGRY = DataTracker.registerData(DubiousPumpkinEntity.class, TrackedDataHandlerRegistry.BOOLEAN);
+    private static final UniformIntProvider ANGER_TIME_RANGE = TimeHelper.betweenSeconds(10, 29);
+    private int angerTime;
+    @Nullable
+    private UUID angryAt;
+
+    public DubiousPumpkinEntity(EntityType<? extends DubiousPumpkinEntity> entityType, World world) {
+        super(entityType, world);
+    }
+
+    @Override
+    protected void initGoals() {
+        super.initGoals();
+        this.goalSelector.add(0, new SwimGoal(this));
+        this.goalSelector.add(1, new DubiousPumpkinEntity.AttackGoal());
+        this.goalSelector.add(5, new WanderAroundGoal(this, 1.0));
+        this.goalSelector.add(6, new LookAtEntityGoal(this, PlayerEntity.class, 6.0F));
+        this.goalSelector.add(7, new LookAroundGoal(this));
+        this.targetSelector.add(1, new DubiousPumpkinEntity.DubiousPumpkinRevengeGoal());
+        this.targetSelector.add(3, new ActiveTargetGoal<>(this, PlayerEntity.class, 10, true, false, this::shouldAngerAt));
+        this.targetSelector.add(4, new ActiveTargetGoal<>(this, ZombieEntity.class, 10, true, true, null));
+        this.targetSelector.add(5, new UniversalAngerGoal<>(this, false));
+    }
+
+    public static DefaultAttributeContainer.Builder createDubiousPumpkinAttributes() {
+        return MobEntity.createMobAttributes().add(EntityAttributes.GENERIC_MAX_HEALTH, 10.0).add(EntityAttributes.GENERIC_FOLLOW_RANGE, 20.0).add(EntityAttributes.GENERIC_MOVEMENT_SPEED, 0.25).add(EntityAttributes.GENERIC_ATTACK_DAMAGE, 1.0);
+    }
+
+    @Override
+    public void readCustomDataFromNbt(NbtCompound nbt) {
+        super.readCustomDataFromNbt(nbt);
+        this.readAngerFromNbt(this.getWorld(), nbt);
+    }
+
+    @Override
+    public void writeCustomDataToNbt(NbtCompound nbt) {
+        super.writeCustomDataToNbt(nbt);
+        this.writeAngerToNbt(nbt);
+    }
+
+    @Override
+    public void chooseRandomAngerTime() {
+        this.setAngerTime(ANGER_TIME_RANGE.get(this.random));
+    }
+
+    @Override
+    public void setAngerTime(int angerTime) {
+        this.angerTime = angerTime;
+    }
+
+    @Override
+    public int getAngerTime() {
+        return this.angerTime;
+    }
+
+    @Override
+    public void setAngryAt(@Nullable UUID angryAt) {
+        this.dataTracker.set(ANGRY, angryAt != null);
+        this.angryAt = angryAt;
+    }
+
+    @Override
+    public void stopAnger() {
+        Angerable.super.stopAnger();
+        this.dataTracker.set(ANGRY, false);
+    }
+
+    @Nullable
+    @Override
+    public UUID getAngryAt() {
+        return this.angryAt;
+    }
+
+    @Override
+    protected SoundEvent getAmbientSound() {
+        return SoundInit.DUBIOUS_PUMPKIN_IDLE_EVENT;
+    }
+
+    @Override
+    protected SoundEvent getHurtSound(DamageSource source) {
+        return SoundInit.DUBIOUS_PUMPKIN_HURT_EVENT;
+    }
+
+    @Override
+    protected SoundEvent getDeathSound() {
+        return SoundInit.DUBIOUS_PUMPKIN_DEATH_EVENT;
+    }
+
+    @Override
+    protected void playStepSound(BlockPos pos, BlockState state) {
+        this.playSound(SoundInit.DUBIOUS_PUMPKIN_STEP_EVENT, 0.15F, 1.0F);
+    }
+
+    @Override
+    protected void initDataTracker(DataTracker.Builder builder) {
+        super.initDataTracker(builder);
+        builder.add(ANGRY, false);
+    }
+
+    @Override
+    public void tick() {
+        super.tick();
+
+        if (!this.getWorld().isClient()) {
+            this.tickAngerLogic((ServerWorld) this.getWorld(), true);
+        }
+    }
+
+    @Override
+    protected float getBaseMovementSpeedMultiplier() {
+        return 0.98F;
+    }
+
+    private class AttackGoal extends MeleeAttackGoal {
+        public AttackGoal() {
+            super(DubiousPumpkinEntity.this, 1.25, true);
+        }
+
+        @Override
+        protected void attack(LivingEntity target) {
+            if (this.canAttack(target)) {
+                this.resetCooldown();
+                this.mob.tryAttack(target);
+            } else if (this.mob.squaredDistanceTo(target) < (target.getWidth() + 3.0F) * (target.getWidth() + 3.0F)) {
+                if (this.isCooledDown()) {
+                    this.resetCooldown();
+                }
+            } else {
+                this.resetCooldown();
+            }
+        }
+
+        @Override
+        public void stop() {
+            super.stop();
+        }
+    }
+
+    private class DubiousPumpkinRevengeGoal extends RevengeGoal {
+        public DubiousPumpkinRevengeGoal() {
+            super(DubiousPumpkinEntity.this);
+        }
+
+        @Override
+        public void start() {
+            super.start();
+            if (DubiousPumpkinEntity.this.isBaby()) {
+                this.callSameTypeForRevenge();
+                this.stop();
+            }
+        }
+
+        @Override
+        protected void setMobEntityTarget(MobEntity mob, LivingEntity target) {
+            if (mob instanceof DubiousPumpkinEntity && !mob.isBaby()) {
+                super.setMobEntityTarget(mob, target);
+            }
+        }
+    }
+
+}